Change id's type from Integer to Long

+++ src/main/java/org/aston/dto/request/LogRequestDTO.java
@@ -8,5 +8,5 @@ import lombok.Data;
 public class LogRequestDTO {
     private String message;
     private String date;
-    private Integer studentId;
+    private Long studentId;
 }
